requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
ansicon==1.89.0
asyncio==3.4.3
attrs==24.2.0
blessed==1.20.0
Brotli==1.1.0
cache==1.0.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
discord.py==2.4.0
editor==1.6.6
frozenlist==1.4.1
idna==3.10
inquirer==3.2.0
jinxed==1.3.0
multidict==6.1.0
mutagen==1.47.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
python-dotenv==1.0.1
readchar==4.2.0
requests==2.32.3
runs==1.2.2
six==1.16.0
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
xmod==1.8.1
yarl==1.11.1
yt-dlp==2024.8.6
